pyproject.toml
[tool.poetry]
name = "yolo-detector"
version = "0.1.0"
description = ""
authors = ["Isma-Ilou Sadou <ismailou.sa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "^1.21.2"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
ultralytics = "^8.0.194"
python-multipart = "^0.0.6"
opencv-python = "^4.8.1.78"
pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
